FSB2.4.1/artifacts/functions/billingFunction/requirements.txt
requests
XlsxWriter
azure-storage-blob==2.1.0
pandas
azure-functions
azure-storage-file==2.1.0
google-cloud-storage==1.36.0
google-auth==1.26.1
numpy
